+++ lib/restclient/payload.rb
@@ -2,17 +2,31 @@ require 'tempfile'
 require 'securerandom'
 require 'stringio'
 
-require 'mime/types'
+begin
+  # Use mime/types/columnar if available, for reduced memory usage
+  require 'mime/types/columnar'
+rescue LoadError
+  require 'mime/types'
+end
 
 module RestClient
   module Payload
     extend self
 
     def generate(params)
-      if params.is_a?(String)
+      if params.is_a?(RestClient::Payload::Base)
+        # pass through Payload objects unchanged
+        params
+      elsif params.is_a?(String)
         Base.new(params)
       elsif params.is_a?(Hash)
         if params.delete(:multipart) == true || has_file?(params)
+          Multipart.new(params)
+        else
+          UrlEncoded.new(params)
+        end
+      elsif params.is_a?(ParamsArray)
+        if _has_file?(params)
           Multipart.new(params)
         else
           UrlEncoded.new(params)
@@ -76,12 +90,20 @@ module RestClient
         @stream.close unless @stream.closed?
       end
 
+      def closed?
+        @stream.closed?
+      end
+
       def to_s_inspect
         to_s.inspect
       end
 
       def short_inspect
-        (size > 500 ? "#{size} byte(s) length" : to_s_inspect)
+        if size && size > 500
+          "#{size} byte(s) length"
+        else
+          to_s_inspect
+        end
       end
 
     end
@@ -98,6 +120,9 @@ module RestClient
           @stream.stat.size
         end
       end
+
+      # TODO (breaks compatibility): ought to use mime_for() to autodetect the
+      # Content-Type for stream objects that have a filename.
 
       alias :length :size
     end