+++ lib/restclient/response.rb
@@ -38,15 +38,25 @@ module RestClient
       "<RestClient::Response #{code.inspect} #{body_truncated(10).inspect}>"
     end
 
-    def self.create(body, net_http_res, request)
+    # Initialize a Response object. Because RestClient::Response is
+    # (unfortunately) a subclass of String for historical reasons,
+    # Response.create is the preferred initializer.
+    #
+    # @param [String, nil] body The response body from the Net::HTTPResponse
+    # @param [Net::HTTPResponse] net_http_res
+    # @param [RestClient::Request] request
+    # @param [Time] start_time
+    def self.create(body, net_http_res, request, start_time=nil)
       result = self.new(body || '')
 
-      result.response_set_vars(net_http_res, request)
+      result.response_set_vars(net_http_res, request, start_time)
       fix_encoding(result)
 
       result
     end
 
+    # Set the String encoding according to the 'Content-Type: charset' header,
+    # if possible.
     def self.fix_encoding(response)
       charset = RestClient::Utils.get_encoding_from_headers(response.headers)
       encoding = nil
@@ -54,8 +64,8 @@ module RestClient
       begin
         encoding = Encoding.find(charset) if charset
       rescue ArgumentError
-        if RestClient.log
-          RestClient.log << "No such encoding: #{charset.inspect}"
+        if response.log
+          response.log << "No such encoding: #{charset.inspect}"
         end
       end